+++ oceanmesh/__init__.py
@@ -1,4 +1,12 @@
-from pyOceanMesh.edgefx import DistanceSizingFunction, Grid
-from pyOceanMesh.geodata import DEM, Geodata, Shoreline
+from oceanmesh.geodata import DEM, Geodata, Shoreline
+from oceanmesh.edgefx import Grid
+from oceanmesh.edgefx import distance_sizing_function
 
-__all__ = ["Grid", "Geodata", "DEM", "Shoreline", "DistanceSizingFunction"]
+__all__ = [
+    "SizeFunction",
+    "Grid",
+    "Geodata",
+    "DEM",
+    "Shoreline",
+    "distance_sizing_function",
+]
